# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Poultry Trading ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (agents)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "400":
          description: Missing or invalid required fields
        "401":
          description: Unauthorized
        "403":
          description: Admin access required
        "500":
          description: Failed to generate upload URL
  /storage/uploads/complete:
    post:
      tags: [storage]
      operationId: completeUpload
      summary: Mark an uploaded object as public-read and get serving URL (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectPath]
              properties:
                objectPath:
                  type: string
                  description: Object path returned by requestUploadUrl (e.g. /objects/uploads/uuid)
      responses:
        "200":
          description: Upload finalized, object is now publicly readable
          content:
            application/json:
              schema:
                type: object
                required: [serveUrl, objectPath]
                properties:
                  serveUrl:
                    type: string
                    description: Public serving URL for the object
                  objectPath:
                    type: string
        "400":
          description: Invalid objectPath
        "401":
          description: Unauthorized
        "403":
          description: Admin access required
        "404":
          description: Object not found — upload may not have completed
        "500":
          description: Failed to finalize upload
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
        "500":
          description: Failed to serve object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        username:
          type: string
        role:
          type: string
          enum: [admin, agent]
        createdAt:
          type: string
      required:
        - id
        - name
        - username
        - role
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, agent]
      required:
        - name
        - username
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, agent]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [chicks, chickens, other]
        unitPrice:
          type: string
        stockQuantity:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - unitPrice
        - stockQuantity
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [chicks, chickens, other]
        unitPrice:
          type: string
        stockQuantity:
          type: integer
        description:
          type: string
      required:
        - name
        - type
        - unitPrice
...
    RequestUploadUrlBody:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    RequestUploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/RequestUploadUrlBody"